moved fetch request to MainContainer
